feat : 버전 맞추기

+++ src/main/java/com/kb_hackathon/plovo/controller/UserController.java
@@ -1,12 +1,16 @@
 package com.kb_hackathon.plovo.controller;
 
+import com.kb_hackathon.plovo.config.S3Uploader;
 import com.kb_hackathon.plovo.dto.RecordRes;
 import com.kb_hackathon.plovo.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
 
 @RestController
 @RequiredArgsConstructor
@@ -16,7 +20,7 @@ public class UserController {
 
     // 기록 페이지 api
     @GetMapping("/auth/record")
-    public RecordRes myRecord(@RequestParam(value = "user_id") int user_id, Authentication authentication) {
+    public RecordRes myRecord(@RequestParam(value = "user_id") Long user_id, Authentication authentication) {
         return userService.myRecord(user_id);
     }
 }
